refactor: move split_observation_features out of livekit.portal into per-package _utils
livekit.portal is transport-layer; lerobot conventions don't belong there.
Each lerobot package now owns its own _utils.py with the helper.
